requirements.txt
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1767524569066/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1751626463503/work
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1767719033/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1691763562544/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_attrs_1764875182/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1764016952863/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1767500808759/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1761202892040/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click==8.3.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coverage @ file:///home/conda/feedstock_root/build_artifacts/coverage_1766951147229/work
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
frozenlist @ file:///home/conda/feedstock_root/build_artifacts/frozenlist_1752167153599/work
greenlet @ file:///home/conda/feedstock_root/build_artifacts/greenlet_1764863560665/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h11_1767729720/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1745602916/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
iniconfig @ file:///home/conda/feedstock_root/build_artifacts/iniconfig_1760831328738/work
jiter @ file:///home/conda/feedstock_root/build_artifacts/jiter_1763006605338/work
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jsonpointer_1765026384/work
langchain @ file:///home/conda/feedstock_root/build_artifacts/langchain_1767933415298/work
langchain-core @ file:///home/conda/feedstock_root/build_artifacts/langchain-core_1767419516866/work
langchain-openai @ file:///home/conda/feedstock_root/build_artifacts/langchain-openai_1767842882033/work
langchain-text-splitters @ file:///home/conda/feedstock_root/build_artifacts/langchain-text-splitters_1756689305564/work
langgraph @ file:///home/conda/feedstock_root/build_artifacts/langgraph_1765620484173/work
langgraph-checkpoint @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_langgraph-checkpoint_1762454552/work
langgraph-cli==0.4.11
langgraph-prebuilt @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_langgraph-prebuilt_1763668305/work
langgraph-sdk @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_langgraph-sdk_1766105007/work
langsmith @ file:///home/conda/feedstock_root/build_artifacts/langsmith_1767930115534/work
loguru @ file:///home/conda/feedstock_root/build_artifacts/loguru_1746634721458/work
multidict @ file:///home/conda/feedstock_root/build_artifacts/multidict_1765460303032/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_numpy_1766373894/work/dist/numpy-2.4.0-cp314-cp314-linux_x86_64.whl#sha256=b85e8eec55fb2b2f3e2ffbbbe557e92140a684ae5dec39b063297fffb9c379a4
openai @ file:///home/conda/feedstock_root/build_artifacts/openai_1766119352666/work
orjson @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_orjson_1765811464/work
ormsgpack @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ormsgpack_1766174725/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pillow @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pillow_1767353193/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pluggy_1764896838/work
propcache @ file:///home/conda/feedstock_root/build_artifacts/propcache_1744524929518/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pydantic_1764434463/work
pydantic_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pydantic-core_1762988995/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytest @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pytest_1765062031/work
pytest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pytest-asyncio_1762797319/work
pytest-cov @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pytest-cov_1757612051/work
pytest-mock @ file:///home/conda/feedstock_root/build_artifacts/pytest-mock_1758101089809/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1761503229/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1758891825788/work
regex @ file:///home/conda/feedstock_root/build_artifacts/regex_1762506938796/work
requests @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_requests_1766926974/work
requests-toolbelt @ file:///home/conda/feedstock_root/build_artifacts/requests-toolbelt_1733734787568/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1762941432086/work
SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/sqlalchemy_1765351722678/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1743640740204/work
tiktoken @ file:///home/conda/feedstock_root/build_artifacts/tiktoken_1764028123207/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1760014679/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
typing-inspection @ file:///home/conda/feedstock_root/build_artifacts/typing-inspection_1764158251755/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1767817748113/work
uuid_utils @ file:///home/conda/feedstock_root/build_artifacts/uuid-utils_1767909998785/work/target/wheels/uuid_utils-0.13.0-cp39-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl#sha256=febb8ae8ff4a1a5c37365bdfe9c9f340604218ebead77c7308f5470e34cadeb1
websockets @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_websockets_1756476367/work/dist
xxhash @ file:///home/conda/feedstock_root/build_artifacts/python-xxhash_1762516172009/work
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1761336872915/work
zstandard==0.25.0
